rename `ConditionsCollapsedSummary` to `ConnectionRowSummary`
