+++ resources/js/components/forms/connections/ConnectionRowSummary.vue
